Crate ws_stream_wasm

Structs

CloseEvent

An event holding information about how/why the connection was closed.

WsMeta

The meta data related to a websocket. Allows access to the methods on the WebSocket API. This is split from the Stream/Sink so you can pass the latter to a combinator whilst continuing to use this API.

WsStream

A futures 0.3 Sink/Stream of WsMessage. Created with WsMeta::connect.

WsStreamIo

A wrapper around WsStream that converts errors into io::Error so that it can be used for io (like AsyncRead/AsyncWrite).

Enums

WsErr

The error type for errors happening in ws_stream_wasm.

WsEvent

Events related to the WebSocket. You can filter like:

WsMessage

Represents a WebSocket Message, after converting from JavaScript type.

WsState

Indicates the state of a Websocket connection. The only state in which it’s valid to send and receive messages is WsState::Open.